webapp-house-finder/requirements.txt
alembic==1.13.1
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.12.3
blinker==1.7.0
Bootstrap-Flask==2.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==38.0.4
dominate==2.9.1
et-xmlfile==1.1.0
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
flask-pyjwt==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
legacy==0.1.7
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.0
python-slugify==8.0.4
pytz==2024.1
soupsieve==2.5
SQLAlchemy==2.0.29
text-unidecode==1.3
typing_extensions==4.11.0
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.2
WTForms==3.1.2
psycopg2-binary